campsite_booking_api/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-cors-headers
psycopg2-binary
django-ratelimit
drf-spectacular
